w2d4

+++ m1-w2d4-introduction-to-classes-exercises/src/main/java/com/techelevator/shoppingcart/ShoppingCart.java
@@ -9,7 +9,7 @@ public class ShoppingCart {
 		
 	}
 	public double getAveragePricePerItem() {
-		return totalAmountOwed / totalNumberOfItems;
+		return totalAmountOwed / (double)totalNumberOfItems;
 		
 	}
